Break up files
